modularize grpc controller and main into main_grpcController
