--- main_grpc.js
@@ -1,2 +1,0 @@
-const grpc = require("@grpc/grpc-js");
-const protoLoader = require("@grpc/proto-loader");
